requirements.txt
Django>=3.0,<4.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
dj-rest-auth==2.2.7
django-extensions==3.2.3
pydot==2.0.0